+++ fastlane_bot/helpers/txhelpers.py
@@ -30,7 +30,7 @@ from web3.types import TxReceipt
 # from fastlane_bot.tools.cpc import ConstantProductCurve
 from fastlane_bot.config import Config
 from fastlane_bot.data.abi import *  # TODO: PRECISE THE IMPORTS or from .. import abi
-from fastlane_bot.utils import num_format, log_format, num_format_float
+from fastlane_bot.utils import num_format, log_format, num_format_float, int_prefix
 
 
 @dataclass
@@ -608,14 +608,12 @@ class TxHelpers:
             if "max fee per gas less than block base fee" in str(e):
                 try:
                     message = str(e)
-                    split1 = message.split("maxFeePerGas: ")[1]
-                    split2 = split1.split(" baseFee: ")
-                    split_baseFee = int(int(split2[1].split(" (supplied gas")[0]))
+                    baseFee = int_prefix(message.split("baseFee: ")[1])
                     transaction = self.construct_contract_function(
                         routes=routes,
                         src_amt=src_amt,
                         src_address=src_address,
-                        gas_price=split_baseFee,
+                        gas_price=baseFee,
                         max_priority=max_priority,
                         nonce=nonce,
                         flashloan_struct=flashloan_struct,
@@ -627,6 +625,7 @@ class TxHelpers:
                     )
                     return None
             else:
+                self.ConfigObj.logger.info(f"gas_price = {gas_price}, max_priority = {max_priority}")
                 self.ConfigObj.logger.warning(
                     f"[helpers.txhelpers.build_transaction_with_gas] (***2***) \n"
                     f"Error when building transaction, this is expected to happen occasionally, discarding. Exception: {e.__class__.__name__} {e}"
@@ -963,13 +962,11 @@ class TxHelpers:
             if "max fee per gas less than block base fee" in str(e):
                 try:
                     message = str(e)
-                    split1 = message.split('maxFeePerGas: ')[1]
-                    split2 = split1.split(' baseFee: ')
-                    split_baseFee = int(int(split2[1].split(" (supplied gas")[0]))
+                    baseFee = int_prefix(message.split("baseFee: ")[1])
                     approve_tx = token_contract.functions.approve(self.arb_contract.address,
                                                                   approval_amount).build_transaction(
                         self.build_tx(
-                            base_gas_price=split_baseFee, max_priority_fee=max_priority, nonce=self.get_nonce()
+                            base_gas_price=baseFee, max_priority_fee=max_priority, nonce=self.get_nonce()
                         )
                     )
                 except Exception as e: