+++ main.py
@@ -14,6 +14,7 @@ check_version_requirements(required_version="6.11.0", package_name="web3")
 import os, sys
 import time
 from typing import List
+from traceback import format_exc
 
 import click
 import pandas as pd
@@ -620,12 +621,6 @@ def run(
                     pool["last_updated_block"] = last_block_queried
                     mgr.pool_data[idx] = pool
 
-            # Log the input passed to function `get_start_block`
-            mgr.cfg.logger.info(
-                "get_start_block input: " +
-                ", ".join([f"({x}, {type(x)})" for x in [alchemy_max_block_fetch, last_block, reorg_delay, replay_from_block]])
-            )
-
             # Get current block number, then adjust to the block number reorg_delay blocks ago to avoid reorgs
             start_block, replay_from_block = get_start_block(
                 alchemy_max_block_fetch, last_block, mgr, reorg_delay, replay_from_block
@@ -848,7 +843,7 @@ def run(
             )
 
         except Exception as e:
-            mgr.cfg.logger.error(f"Error in main loop: {e}")
+            mgr.cfg.logger.error(f"Error in main loop: {format_exc()}")
             mgr.cfg.logger.error(
                 f"[main] Error in main loop: {e}. Continuing... "
                 f"Please report this error to the Fastlane Telegram channel if it persists."